requirements.txt
flask
datasets
pyyaml
python-dotenv
Pillow
tqdm
qdrant-client
numpy